+++ sdk/http/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @outputai/http
 
+## 0.8.1
+
+### Patch Changes
+
+- Updated dependencies [aa8ed5e]
+  - @outputai/core@0.8.1
+
 ## 0.8.0
 
 ### Patch Changes